JE2-251 Added function getUsersInitials
